pwsh: update windows alias
- disable removing ls and rm alias on non-windows system

+++ dot_config/powershell/alias.ps1
@@ -1,4 +1,6 @@
 # Write-Output "Module alias loaded"
 Set-Alias -Name cd -Value __zoxide_z -Option AllScope -Scope Global -Force
-Remove-Alias -Name ls
-Remove-Alias -Name rm
+if ($env.OS -eq "Windows_NT") {
+	Remove-Alias -Name ls
+	Remove-Alias -Name rm
+}
